feat: add validity check for valid_lens in DotProductAttention and create test header for attention functions
